Consider upstream k8s.io/kubectl/pkg/drain to implement drain
* Our original drain implementation may be preferrable
* The upstream implementation is atrocious and brings in a bunch
of indirect dependencies and bloat
  * RunNodeDrain requires a drain.Helper but has no constructor
    function, no config struct, and a bunch of exported fields
    (some are required to avoid panics, others the defaults are fine,
    you get to guess!)
  * RunNodeDrain acts on a node name string, while RunCordonOrUncordon
  (ridiculous name) needs one to create a skeleton v1core.Node struct
  for no reason
  * Littered with anti-patterns (factories, util everything, exported
  types that should be internal)
  * Honestly, this package wasn't designed for actual humans to use
